refactor(users): move useCurrentUser to lib/users/hooks + expose isLoading
Named export, "use client" directive, isLoading added to return shape.
